Work on NO_CYCLECNT == 1
The cycle counter is optional. So far it is working as expected in a
highlevel implementation, when the cyclecounter is really
present. This commit corrected an error seen when the cyclecounter is
not present, for the highlevel code.

The cyclecounter is also connected to startup conditions of midgetv,
this is what I look at now.